requirements.txt
hydra-core==1.3.2
matplotlib==3.8.2
numpy==1.26.2
omegaconf==2.3.0
open3d==0.17.0
scipy==1.11.4
torch==2.2.0.dev20230929
tqdm==4.66.1
transformers==4.35.2
